requirements.txt
# AutoReel Agent - Dependencies

# Transcription
openai-whisper>=20231117
transformers>=4.36.0
torch>=2.1.0

# Video Processing
ffmpeg-python>=0.2.0
yt-dlp>=2024.1.0

# LLM / AI
google-generativeai>=0.8.0

# Data Models & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0

# Hebrew NLP
regex>=2023.0

# Subtitle Processing
pysrt>=1.1.2

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0          # Beautiful terminal output
tqdm>=4.66.0          # Progress bars
click>=8.1.0          # CLI framework

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
